recsys_webapp/requirements.txt
annoy==1.17.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.4
